refactor: Generalize and relocation Fields API Factory.
