--- src/Framework/FieldsAPI/Factory/README.md
@@ -1,5 +1,0 @@
-# Factories
-
-## Factroy vs Static Constructors
-
-Using a Factory class, over a static constructor for each type, has the benefit of making field types discoverable through editor auto-completion.
