+++ src/Framework/FieldsAPI/Hidden.php
@@ -4,6 +4,8 @@ namespace Give\Framework\FieldsAPI;
 
 class Hidden extends Field {
 
+	use Concerns\HasLabel;
+	use Concerns\HasEmailTag;
 	use Concerns\ShowInReceipt;
 	use Concerns\StoreAsMeta;
